Welcome Message Fix
Jolene would post in the specified welcome message channel if anyone joined any server Jolene was in. Now, it will assign the value from the member's perspective instead of the bots. This means if someone joins on a different server, join_notif_channel might be defined/null, and so a check was implemented to catch that. If there's no value, then nothing gets posted.